+++ CHANGES.md
@@ -13,6 +13,7 @@ Change Log
   * This supports `zip` files through add local/remote data
   * Supports over 40 formats - including Collada, obj, Blender, DXF - [full list](https://github.com/assimp/assimp/blob/master/doc/Fileformats.md)
 * Add `description` to `getDataType` - this will be displayed between Step 1 and Step 2
+* Add warning message to `GltfMixin` when showing in 2D mode (Leaflet)
 * [The next improvement]
 
 #### 8.2.8 - 2022-07-04